requirements.txt
certifi==2016.8.31
future==0.15.2
python-telegram-bot==5.0.0
requests==2.11.1
six==1.10.0
spotipy==2.3.8
urllib3==1.17
wheel==0.26.0
